+++ Image.h
@@ -40,19 +40,18 @@ struct Image {
     Image colorMaskNew(float r, float g, float b) const;
     Image colorMaskNew(uint8_t r, uint8_t g, uint8_t b) const;
     Image &overlay(const Image &source, int x, int y);
-    Image resizeFastNew(uint16_t rw, uint16_t rh) const;
-    Image cropNew(uint16_t cx, uint16_t cy, uint16_t cw, uint16_t ch) const;
+    Image resizeFastNew(int rw, int rh) const;
+    Image cropNew(int cx, int cy, int cw, int ch) const;
 
     Image quadifyFrameBW(std::map<std::pair<int, int>, Image> &resizedAmogi) const;
-    void subdivideBW(uint16_t sx, uint16_t sy, uint16_t sw, uint16_t sh, Image &frame,
+    void subdivideBW(int sx, int sy, int sw, int sh, Image &frame,
                      std::map<std::pair<int, int>, Image> &resizedAmogi) const;
-    std::tuple<bool, uint8_t> subdivideCheckBW(uint16_t sx, uint16_t sy, uint16_t sw, uint16_t sh) const;
+    std::tuple<bool, uint8_t> subdivideCheckBW(int sx, int sy, int sw, int sh) const;
 
     Image quadifyFrameRGB(std::map<std::pair<int, int>, Image> &resizedAmogi) const;
-    void subdivideRGB(uint16_t sx, uint16_t sy, uint16_t sw, uint16_t sh, Image &frameRGB,
+    void subdivideRGB(int sx, int sy, int sw, int sh, Image &frameRGB,
                       std::map<std::pair<int, int>, Image> &resizedAmogi) const;
-    std::tuple<bool, uint8_t, uint8_t, uint8_t> subdivideCheckRGB(uint16_t sx, uint16_t sy, uint16_t sw,
-                                                                  uint16_t sh) const;
+    std::tuple<bool, uint8_t, uint8_t, uint8_t> subdivideCheckRGB(int sx, int sy, int sw, int sh) const;
 
     std::map<std::pair<int, int>, Image> preloadResized(int sw, int sh) const;
     void subdivideValues(int sx, int sy, int sw, int sh, std::map<std::pair<int, int>, Image> &image_map) const;
